feat: correct format for ft value display (#124)

+++ apps/explorer/priv/repo/migrations/20230210192835_create_forward_transfers.exs
@@ -4,7 +4,7 @@ defmodule Explorer.Repo.Migrations.CreateForwardTransfers do
   def change do
     create table(:forward_transfers, primary_key: true) do
       add(:to_address_hash, :string, null: false)
-      add(:value, :bigint, null: false)
+      add(:value, :numeric, precision: 100, null: false)
       add(:block_number, :bigint, null: false)
       timestamps(null: false, type: :utc_datetime_usec)
     end
